// front/src/types/companyType.ts
export interface PhotosSample {
  photo_id: string;
  photo_url: string;
  photo_url_large: string;
  video_thumbnail_url: any;
  latitude: number;
  longitude: number;
  type: string;
  photo_datetime_utc: string;
  photo_timestamp: number;
  file: any;
}